bdr: fix: use correct datatypes in various loops
This avoids compiler warnings for signed vs unsigned comparison.

+++ contrib/bdr/bdr_seq.c
@@ -526,7 +526,7 @@ bdr_sequencer_shmem_init(int nnodes, int sequencers)
 void
 bdr_sequencer_wakeup(void)
 {
-	int off;
+	size_t off;
 	BdrSequencerSlot *slot;
